fix(autopilot): EXTRACT returns numeric, and it aborted every brain cycle
Root cause of the degradation that started at 06:25 today and stopped the brain
recording a single decision for two hours — including against the 412 live
opportunities imported in dc6723f.

`growth_intelligence/worker_signals.rs` selects

    EXTRACT(EPOCH FROM (oa.created_at - action.created_at)) / 3600.0

into an `f64`. PostgreSQL 14 changed `EXTRACT` to return `numeric` where it used
to return `double precision`, and dividing by a numeric literal keeps it numeric,
so both columns arrived as NUMERIC and the decode failed:

    error occurred while decoding column 2: mismatched types;
    Rust type `f64` (as SQL type `FLOAT8`) is not compatible with SQL type `NUMERIC`

The whole evaluation phase aborts on it, which is why `decisions_recorded` was 0
on every cycle while the rest of the cycle carried on and reported `degraded`.

The timing is the interesting part. That query reads operator approvals, so it
returned zero rows for as long as the loop had never produced anything to approve
— and started failing the moment it did, nine decisions in, the first time the
brain had ever got that far. It compiled, linted and passed 1684 tests, because
no test executes it with a row in it. Exactly the standing risk CLAUDE.md names
for the runtime-query choice, and exactly what "machinery never exercised under
load" means.

Also fixes the same mistake in `decisions/opportunity_reads.rs`, where two
`SUM(...) OVER (...)` windows over `*_minor` bigint columns return NUMERIC into
`i64` fields. Latent — it needs one `viryaos_promotion_campaign_states` row, and
production has none — but it is the identical defect one table over.

Finding it took hours of reading that should not have been necessary: a decode
error names the column index and the two types, and nothing about which of ~1700
runtime queries produced it. A backtrace was added at the `map_sqlx` funnel
(7f7d3e0) and turned on in production (b6db745); it came back all `<unknown>`
because the release binary is stripped. What actually found it was asking
PostgreSQL: `PREPARE` every query and read
`pg_prepared_statements.result_types`.

So that is now the gate. `scripts/sql-result-types.py` sweeps all 736
row-returning literals, prepares 674 of them and refuses any NUMERIC output
column, naming the likely expression — EXTRACT, SUM over bigint, AVG over
integer. Nothing in this schema stores `numeric` except
`merch_coupons.discount_percent`, so a NUMERIC result is always an accident.
Verified it bites: reverting one cast fails it with the file, line, column index
and the fix. Added to `just policy-checks`; skips without a local database, like
the cross-repo gates.

CLAUDE.md's policy-script count goes 23 to 24, which its own accuracy gate
caught.

+++ crates/crowdrelay-infra/src/autopilot/decisions/opportunity_reads.rs
@@ -248,12 +248,20 @@ macro_rules! decision_opportunity_reads {
                     state.maximum_daily_budget_minor,
                     state.spend_last_7d_minor,
                     state.attributed_revenue_last_7d_minor,
+                    -- `::bigint` on both. `SUM` over a `bigint` column returns
+                    -- NUMERIC in PostgreSQL, and both of these are `*_minor`
+                    -- bigints, so without the cast the two columns arrive as
+                    -- NUMERIC and the `i64` fields they decode into fail. Latent
+                    -- rather than live: it needs one `viryaos_promotion_campaign_states`
+                    -- row to fire, and production has none yet. Its sibling in
+                    -- `growth_intelligence/worker_signals.rs` was the same mistake
+                    -- and did fire, aborting every autopilot cycle for two hours.
                     SUM(state.current_daily_budget_minor) OVER (
                         PARTITION BY state.workspace_id, state.currency
-                    ) AS workspace_daily_budget_minor,
+                    )::bigint AS workspace_daily_budget_minor,
                     SUM(state.spend_month_to_date_minor) OVER (
                         PARTITION BY state.workspace_id, state.currency
-                    ) AS workspace_spend_month_to_date_minor,
+                    )::bigint AS workspace_spend_month_to_date_minor,
                     guardrail.maximum_total_daily_budget_minor AS workspace_maximum_daily_budget_minor,
                     guardrail.maximum_monthly_spend_minor AS workspace_maximum_monthly_spend_minor,
                     CASE
